fix: webpack 5 support
