+++ src/utils.js
@@ -49,7 +49,26 @@ function compareModulesByIdentifier(a, b) {
   return compareIds(a.identifier(), b.identifier());
 }
 
+const initializeCache = new WeakMap();
+
+function shared(webpack, initializer) {
+  const cacheEntry = initializeCache.get(webpack);
+
+  // eslint-disable-next-line no-undefined
+  if (cacheEntry !== undefined) {
+    return cacheEntry;
+  }
+
+  const constructors = initializer(webpack);
+  const result = { ...constructors };
+
+  initializeCache.set(webpack, result);
+
+  return result;
+}
+
 export {
+  shared,
   MODULE_TYPE,
   findModuleById,
   evalModuleCode,
