+++ test/TestCache.test.js
@@ -43,28 +43,22 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
-          expect(
-            Array.from(stats.compilation.emittedAssets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          `);
+          expect(Array.from(stats.compilation.emittedAssets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(stats.compilation.warnings).toHaveLength(0);
           expect(stats.compilation.errors).toHaveLength(0);
 
@@ -94,17 +88,14 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(
             Array.from(stats.compilation.emittedAssets).sort()
           ).toMatchInlineSnapshot(`Array []`);
@@ -156,28 +147,22 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
-          expect(
-            Array.from(stats.compilation.emittedAssets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          `);
+          expect(Array.from(stats.compilation.emittedAssets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(stats.compilation.warnings).toHaveLength(0);
           expect(stats.compilation.errors).toHaveLength(0);
 
@@ -207,17 +192,14 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(
             Array.from(stats.compilation.emittedAssets).sort()
           ).toMatchInlineSnapshot(`Array []`);
@@ -274,16 +256,13 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
             ]
-          `
-          );
+          `);
           expect(Array.from(stats.compilation.emittedAssets).sort())
             .toMatchInlineSnapshot(`
             Array [
@@ -323,16 +302,13 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
             ]
-          `
-          );
+          `);
           expect(
             Array.from(stats.compilation.emittedAssets).sort()
           ).toMatchInlineSnapshot(`Array []`);
@@ -389,28 +365,22 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
-          expect(
-            Array.from(stats.compilation.emittedAssets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          `);
+          expect(Array.from(stats.compilation.emittedAssets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(stats.compilation.warnings).toHaveLength(0);
           expect(stats.compilation.errors).toHaveLength(0);
 
@@ -443,17 +413,14 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(
             Array.from(stats.compilation.emittedAssets).sort()
           ).toMatchInlineSnapshot(`Array []`);
@@ -510,28 +477,22 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
-          expect(
-            Array.from(stats.compilation.emittedAssets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          `);
+          expect(Array.from(stats.compilation.emittedAssets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(stats.compilation.warnings).toHaveLength(0);
           expect(stats.compilation.errors).toHaveLength(0);
 
@@ -564,17 +525,14 @@ describe('TestCache', () => {
             return;
           }
 
-          expect(
-            Object.keys(stats.compilation.assets).sort()
-          ).toMatchInlineSnapshot(
-            `
+          expect(Object.keys(stats.compilation.assets).sort())
+            .toMatchInlineSnapshot(`
             Array [
               "main.css",
               "main.js",
               "static/react.svg",
             ]
-          `
-          );
+          `);
           expect(
             Array.from(stats.compilation.emittedAssets).sort()
           ).toMatchInlineSnapshot(`Array []`);
